backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
jieba==0.42.1
python-docx==1.1.0
numpy==1.26.3
